# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fraud Detection System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
  /notifications/unread-count:
    get:
      operationId: getUnreadCount
      tags: [notifications]
      summary: Get unread notification count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                required: [count]
  /notifications/read-all:
    post:
      operationId: markAllRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppNotification"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all fraud cases
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, under_review, resolved, dismissed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FraudCase"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCase"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a single case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCase"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseBody"
      responses:
        "200":
          description: Case updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudCase"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Case deleted
  /cases/{id}/notes:
    get:
      operationId: listCaseNotes
      tags: [cases]
      summary: List notes for a case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseNote"
    post:
      operationId: addCaseNote
      tags: [cases]
      summary: Add a note to a case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddNoteBody"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseNote"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    AppNotification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
          nullable: true
        caseId:
          type: integer
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - title
        - isRead
        - createdAt
    FraudCase:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, under_review, resolved, dismissed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        alertId:
          type: integer
          nullable: true
        transactionRef:
          type: string
          nullable: true
        merchantName:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        location:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        assigneeName:
          type: string
          nullable: true
        createdBy:
          type: integer
          nullable: true
        noteCount:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        resolvedAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt
    CaseNote:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        content:
          type: string
        createdBy:
          type: integer
          nullable: true
        authorName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - caseId
        - content
        - createdAt
    CreateCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        alertId:
          type: integer
        transactionRef:
          type: string
        merchantName:
          type: string
        amount:
          type: number
        location:
          type: string
        assignedTo:
          type: integer
      required:
        - title
    UpdateCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, resolved, dismissed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedTo:
          type: integer
          nullable: true
    AddNoteBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...